@@ -1,46 +1,71 @@
 require 'spec_helper'
 
-feature "Update Menu" do
-	
-# 	given(:menu) do
-# 		FactorysGirl.create(:menu)
-# 	end
+feature 'only shows relevant actions to non-admin' do
+    
+  let(:menu) {FactoryGirl.create(:chili_pepper_menu)}
+  let(:section) { FactoryGirl.build(:chili_pepper_section, :menu => nil) }
 
-# 	scenario "shoud not allow me to visit the edit page without being an admin" do
-# 	  visit edit_menu_path(menu)
-# 	  page.should have_content('Salut')
-# 	end
+  scenario 'does not allow non-admin to visit the new or edit pages' do
+    visit chili_pepper.new_menu_path
+    expect(page).not_to have_content('New')
+    visit chili_pepper.edit_menu_path(menu)
+    expect(page).not_to have_content('Edit')
+  end
 
-# 	scenario "should allow me to update the proper menu when logged in as admin" do
-# 		admin = FactoryGirl.create(:admin)
-# 	  	visit '/admins/sign_in' 
-# 	  	fill_in "Email", with: admin.email
-# 	  	fill_in "Password", with: admin.password
-# 	  	click_button 'Log in'
-		
-# 		visit edit_menu_path(menu)
-# 		page.should have_content "Edit #{menu.name}"
-# 		fill_in "Name", with: 'Babou'
-# 	    click_button "Update Menu"
-# 	    page.should have_content('Babou')
-# 	end		
+  scenario "menu page redirects to section page if section exists" do
+    menu.sections << section
+    visit chili_pepper.menu_path(menu)
+    expect(current_path).to eq(chili_pepper.menu_section_path(menu, section))
+  end
 
-# end
+  scenario "menu page just renders blank menu page if section doesn't exists" do
+    visit chili_pepper.menu_path(menu)
+    expect(current_path).to eq(chili_pepper.menu_path(menu))
+  end
 
-# feature "Display Properly Menu" do
-# 	given(:menu) do
-# 		FactoryGirl.create(:menu)
-# 	end
+end
 
-# 	scenario "displays the proper menu" do
-# 		dish_1 = FactoryGirl.create(:dish, :column => 1, :menu_id => menu.id)
-# 		dish_2 = FactoryGirl.create(:dish, :column => 2, :menu_id => menu.id)
-# 		visit menu_path(menu)
-# 		page.should have_content menu.name
-# 		page.should have_selector("#column_1 #dish_#{dish_1.id}")
-# 		page.should have_selector("#column_2 #dish_#{dish_2.id}")
+feature 'let admin manage menus' do
 
+  let(:menu) {FactoryGirl.create(:chili_pepper_menu)}
 
-# 	end
+  before :each do
+    admin = FactoryGirl.create(:chili_pepper_admin)
+    visit 'chili_pepper/admins/sign_in' 
+    fill_in 'Email', with: admin.email
+    fill_in 'Password', with: admin.password
+    click_button 'Sign in'
+  end
+
+  scenario 'should allow me to add a menu when logged in as admin' do
+      visit '/chili_pepper/menus/new'
+      expect(page).to have_content 'New Menu'
+      fill_in 'Name', with: 'Babou'
+      click_button 'Create Menu'
+      expect(current_path).to eq(chili_pepper.menu_path('babou'))
+      expect(page).to have_content('Babou')
+  end
+
+  scenario 'allows to edit menu' do
+    visit chili_pepper.edit_menu_path(menu)
+    fill_in 'Name', with: 'PoteMenu Edited'
+    click_button 'Update Menu'
+    expect(current_path).to eq(chili_pepper.menu_path(menu))
+    expect(page).to have_content('PoteMenu Edited')
+  end
+
+  scenario 'allows to delete menu' do
+    visit chili_pepper.menu_path(menu)
+    click_link 'Delete Menu'
+   visit chili_pepper.menus_path
+    expect(page).not_to have_content('A La Carte')
+    expect(page).to have_content('New Menu')
+  end
+
+  scenario 'allows visit index of menus' do
+    visit chili_pepper.menus_path
+    visit chili_pepper.menu_path(menu)
+    expect(page).to have_content('A La Carte')
+  end
 
 end
