@@ -1,4 +1,4 @@
 module ChiliPepper
-  class ApplicationController < ActionController::Base
+  class ApplicationController < ::ApplicationController
   end
 end
