@@ -4,22 +4,90 @@ require 'rubygems'
 require 'spork'
 require 'factory_girl'
 require 'capybara/rspec'
+require 'capybara/poltergeist'
+Capybara.javascript_driver = :poltergeist
 # require "paperclip/matchers"
 #uncomment the following line to use spork with the debugger
 #require 'spork/ext/ruby-debug'
+
+# This file is copied to spec/ when you run 'rails generate rspec:install'
+
+
+# Requires supporting ruby files with custom matchers and macros, etc,
+# in spec/support/ and its subdirectories.
+
+# Checks for pending migrations before tests are run.
+# If you are not using ActiveRecord, you can remove this line.
 
 Spork.prefork do
   ENV["RAILS_ENV"] ||= 'test'
   require File.expand_path("../../spec/test_app/config/environment", __FILE__)
   require 'rspec/rails'
-  require 'rspec/autorun'
+  require 'draper/test/rspec_integration'
+  require 'database_cleaner'
 
   Dir[Rails.root.join("spec/support/**/*.rb")].each {|f| require f}
 
+
   RSpec.configure do |config|
+
+    
+
     config.fixture_path = "#{::Rails.root}/spec/fixtures"
     config.use_transactional_fixtures = true
     config.infer_base_class_for_anonymous_controllers = false
+    config.include Devise::TestHelpers, type: :controller
+    # ## Mock Framework
+    #
+    # If you prefer to use mocha, flexmock or RR, uncomment the appropriate line:
+    #
+    # config.mock_with :mocha
+    # config.mock_with :flexmock
+    # config.mock_with :rr
+
+    # Remove this line if you're not using ActiveRecord or ActiveRecord fixtures
+    # config.fixture_path = "#{::Rails.root}/spec/fixtures"
+
+    config.before(:suite) do
+       DatabaseCleaner.clean_with(:truncation)
+     end
+    
+     config.before(:each) do
+       DatabaseCleaner.strategy = :transaction
+     end
+    
+     config.before(:each, js: true) do
+       DatabaseCleaner.strategy = :truncation
+     end
+    
+     config.before(:each) do
+       DatabaseCleaner.start
+     end
+    
+     config.after(:each) do
+       DatabaseCleaner.clean
+     end
+
+    config.after(:suite) do
+      FileUtils.rm_rf(Dir["#{::Rails.root}/spec/test_files/"])
+    end
+
+    # If you're not using ActiveRecord, or you'd prefer not to run each of your
+    # examples within a transaction, remove the following line or assign false
+    # instead of true.
+    config.use_transactional_fixtures = false
+
+    # If true, the base class of anonymous controllers will be inferred
+    # automatically. This will be the default behavior in future versions of
+    # rspec-rails.
+    config.infer_base_class_for_anonymous_controllers = false
+
+    # Run specs in random order to surface order dependencies. If you find an
+    # order dependency and want to debug it, you can fix the order by providing
+    # the seed, which is printed after each run.
+    #     --seed 1234
+    config.order = "random"
+
   end
 end
 
@@ -60,48 +128,7 @@ end
 
 
 
-# This file is copied to spec/ when you run 'rails generate rspec:install'
-ENV["RAILS_ENV"] ||= 'test'
-require File.expand_path("../test_app/config/environment", __FILE__)
-require 'rspec/rails'
-require 'rspec/autorun'
 
-# Requires supporting ruby files with custom matchers and macros, etc,
-# in spec/support/ and its subdirectories.
-Dir[Rails.root.join("spec/support/**/*.rb")].each { |f| require f }
-
-# Checks for pending migrations before tests are run.
-# If you are not using ActiveRecord, you can remove this line.
-ActiveRecord::Migration.check_pending! if defined?(ActiveRecord::Migration)
-
-RSpec.configure do |config|
-  # ## Mock Framework
-  #
-  # If you prefer to use mocha, flexmock or RR, uncomment the appropriate line:
-  #
-  # config.mock_with :mocha
-  # config.mock_with :flexmock
-  # config.mock_with :rr
-
-  # Remove this line if you're not using ActiveRecord or ActiveRecord fixtures
-  config.fixture_path = "#{::Rails.root}/spec/fixtures"
-
-  # If you're not using ActiveRecord, or you'd prefer not to run each of your
-  # examples within a transaction, remove the following line or assign false
-  # instead of true.
-  config.use_transactional_fixtures = true
-
-  # If true, the base class of anonymous controllers will be inferred
-  # automatically. This will be the default behavior in future versions of
-  # rspec-rails.
-  config.infer_base_class_for_anonymous_controllers = false
-
-  # Run specs in random order to surface order dependencies. If you find an
-  # order dependency and want to debug it, you can fix the order by providing
-  # the seed, which is printed after each run.
-  #     --seed 1234
-  config.order = "random"
-end
 
 Spork.each_run do
   # This code will be run each time you run your specs.
